[Exchange Oracle] Autofix F541: f-string-missing-placeholders
Checks for f-strings that do not contain any placeholder expressions.

f-strings are a convenient way to format strings, but they are not
necessary if there are no placeholder expressions to format. In this
case, a regular string should be used instead, as an f-string without
placeholders can be confusing for readers, who may expect such a
placeholder to be present.

An f-string without any placeholders could also indicate that the
author forgot to add a placeholder expression.

```python
f"Hello, world!"
```

Use instead:
```python
"Hello, world!"
```

**Note:** to maintain compatibility with PyFlakes, this rule only flags
f-strings that are part of an implicit concatenation if _none_ of the
f-string segments contain placeholder expressions.

For example:

```python
(
    f"Hello,"
    f" {name}!"
)

(
    f"Hello,"
    f" World!"
)
```

See [#10885](https://github.com/astral-sh/ruff/issues/10885) for more.

- [PEP 498](https://www.python.org/dev/peps/pep-0498/)