[Exchange Oracle] Autofix F841: unused-variable
Checks for the presence of unused variables in function scopes.

A variable that is defined but not used is likely a mistake, and should
be removed to avoid confusion.

If a variable is intentionally defined-but-not-used, it should be
prefixed with an underscore, or some other value that adheres to the
[`lint.dummy-variable-rgx`] pattern.

Under [preview mode](https://docs.astral.sh/ruff/preview), this rule also
triggers on unused unpacked assignments (for example, `x, y = foo()`).

```python
def foo():
    x = 1
    y = 2
    return x
```

Use instead:
```python
def foo():
    x = 1
    return x
```

- `lint.dummy-variable-rgx`