[Exchange Oracle] Unsafe autofix C408: unnecessary-collection-call
Checks for unnecessary `dict`, `list` or `tuple` calls that can be
rewritten as empty literals.

It's unnecessary to call, e.g., `dict()` as opposed to using an empty
literal (`{}`). The former is slower because the name `dict` must be
looked up in the global scope in case it has been rebound.

```python
dict()
dict(a=1, b=2)
list()
tuple()
```

Use instead:
```python
{}
{"a": 1, "b": 2}
[]
()
```

This rule's fix is marked as unsafe, as it may occasionally drop comments
when rewriting the call. In most cases, though, comments will be preserved.

- `lint.flake8-comprehensions.allow-dict-calls-with-keyword-arguments`