[Exchange Oracle] Unsafe autofix SIM103: needless-bool
Checks for `if` statements that can be replaced with `bool`.

`if` statements that return `True` for a truthy condition and `False` for
a falsey condition can be replaced with boolean casts.

Given:
```python
if x > 0:
    return True
else:
    return False
```

Use instead:
```python
return x > 0
```

Or, given:
```python
if x > 0:
    return True
return False
```

Use instead:
```python
return x > 0
```

- [Python documentation: Truth Value Testing](https://docs.python.org/3/library/stdtypes.html#truth-value-testing)