[Exchange Oracle] Unsafe autofix SIM108: if-else-block-instead-of-if-exp
Check for `if`-`else`-blocks that can be replaced with a ternary operator.
Moreover, in [preview], check if these ternary expressions can be
further simplified to binary expressions.

`if`-`else`-blocks that assign a value to a variable in both branches can
be expressed more concisely by using a ternary or binary operator.

```python
if foo:
    bar = x
else:
    bar = y
```

Use instead:
```python
bar = x if foo else y
```

Or, in [preview]:

```python
if cond:
    z = cond
else:
    z = other_cond
```

Use instead:

```python
z = cond or other_cond
```

- [Python documentation: Conditional expressions](https://docs.python.org/3/reference/expressions.html#conditional-expressions)

[preview]: https://docs.astral.sh/ruff/preview/