[Exchange Oracle] Unsafe autofix TRY201: verbose-raise
Checks for needless exception names in `raise` statements.

It's redundant to specify the exception name in a `raise` statement if the
exception is being re-raised.

```python
def foo():
    try:
        ...
    except ValueError as exc:
        raise exc
```

Use instead:
```python
def foo():
    try:
        ...
    except ValueError:
        raise
```

This rule's fix is marked as unsafe, as it doesn't properly handle bound
exceptions that are shadowed between the `except` and `raise` statements.