[Recording Oracle] Unsafe autofix B007: unused-loop-control-variable
## What it does
Checks for unused variables in loops (e.g., `for` and `while` statements).

## Why is this bad?
Defining a variable in a loop statement that is never used can confuse
readers.

If the variable is intended to be unused (e.g., to facilitate
destructuring of a tuple or other object), prefix it with an underscore
to indicate the intent. Otherwise, remove the variable entirely.

## Example
```python
for i, j in foo:
    bar(i)
```

Use instead:
```python
for i, _j in foo:
    bar(i)
```

## References
- [PEP 8: Naming Conventions](https://peps.python.org/pep-0008/#naming-conventions)